Split handler tests and add edge case coverage
- Split handler_test.go into separate files matching source structure
- Add edge case tests for malicious content, Unicode, max length validation
- Update portfolio-common to v0.32.0 for CloseDB helper
- Add RabbitMQ config to docker-compose and .env.example
