+++ .env.example
@@ -2,7 +2,7 @@
 PORT=8086
 ENVIRONMENT=development
 
-# Database Configuration
+# Database Configuration (dev only - use secrets manager in production)
 DB_HOST=localhost
 DB_PORT=5432
 DB_USER=portfolio_messaging
@@ -10,17 +10,17 @@ DB_PASSWORD=portfolio_messaging_dev_pass
 DB_NAME=portfolio
 DB_SSL_MODE=disable
 
-# JWT Authentication (required, min 32 characters)
+# JWT Authentication (dev only - use strong secret in production)
 JWT_SECRET=your-secret-key-at-least-32-characters
 
 # CORS Configuration (required)
 ALLOWED_ORIGINS=http://localhost:3000
 
-# RabbitMQ Configuration (required for message queue)
+# RabbitMQ Configuration (dev only - use strong credentials in production)
 RABBITMQ_HOST=localhost
 RABBITMQ_PORT=5672
-RABBITMQ_USER=guest
-RABBITMQ_PASSWORD=guest
+RABBITMQ_USER=messaging
+RABBITMQ_PASSWORD=messaging_dev_pass
 RABBITMQ_EXCHANGE=contact_messages
 RABBITMQ_QUEUE=contact_messages
 RABBITMQ_RETRY_DELAYS=1m,5m,30m,2h,12h
