+++ TESTING.md
@@ -30,16 +30,54 @@ go test -v -run Recipient ./internal/handlers/
 
 ## Test Files
 
-**`handler_test.go`** - 43 tests
+Tests are organized by source file, following Go conventions:
 
-| Category | Tests | Coverage |
-|----------|-------|----------|
-| Health Check | 1 | HealthCheck endpoint |
-| Contact Messages | 11 | Create, GetAll, GetByID + error cases |
-| Recipients | 20 | GetAll, GetByID, Create, Update, Delete + errors |
-| Context Propagation | 3 | Verifies context with sentinel value |
-| ID Validation | 2 | Table-driven invalid ID format tests (16 subtests) |
-| Constructor | 1 | Handler initialization |
+| File | Tests | Description |
+|------|-------|-------------|
+| `handler_test.go` | 1 | Constructor |
+| `health_test.go` | 1 | Health check endpoint |
+| `contact_message_test.go` | 20 | Contact message CRUD + edge cases |
+| `recipient_test.go` | 20 | Recipient CRUD operations |
+| `mocks_test.go` | - | Mocks and shared test utilities |
+
+### Test Breakdown
+
+**`handler_test.go`** (1 test)
+
+- Handler constructor initialization
+
+**`health_test.go`** (1 test)
+
+- Health check endpoint success
+
+**`contact_message_test.go`** (20 tests)
+
+- CreateContactMessage: success, spam, publish error, validation, repo error, JSON
+- CreateContactMessage edge cases: Unicode, max length, empty honeypot, malicious
+- GetContactMessages: success, empty list, repository error
+- GetContactMessage: success, invalid ID, not found, repository error
+- Context propagation tests
+- Invalid ID format validation (table-driven)
+
+**`recipient_test.go`** (20 tests)
+
+- GetRecipients: success, empty list, repository error
+- GetRecipient: success, invalid ID, not found, repository error
+- CreateRecipient: success, isActive=false, validation, email, repo error, JSON
+- UpdateRecipient: success, partial, invalid ID, not found, validation, repo error
+- DeleteRecipient: success, invalid ID, not found, repository error
+- Context propagation tests
+- Invalid ID format validation (table-driven)
+
+**`mocks_test.go`** (shared utilities)
+
+- `mockRepository` - configurable mock implementing Repository interface
+- `mockPublisher` - configurable mock implementing Publisher interface
+- `setupTestRouter()` - creates Gin router in test mode
+- `performRequest()` - executes HTTP request against router
+- `createTestContactMessage()` / `createTestContactMessages()` - factory functions
+- `createTestRecipient()` / `createTestRecipients()` - factory functions
+- `ctxKey` - context key type for propagation tests
 
 ## Key Testing Patterns
 
@@ -51,6 +89,17 @@ mockRepo := &mockRepository{
         ctx context.Context,
         msg *models.ContactMessage,
     ) error {
+        return nil
+    },
+}
+```
+
+**Mock Publisher**: Verifies queue integration
+
+```go
+mockPub := &mockPublisher{
+    publishFunc: func(ctx context.Context, msg interface{}) error {
+        // Capture published message for assertions
         return nil
     },
 }
@@ -81,6 +130,7 @@ recipient := createTestRecipient()
 - Sets correct HTTP status
 - Sets Location header on create
 - Spam detection returns success but doesn't save
+- Queue publish is called with correct message event
 
 ### Error Cases
 
@@ -90,6 +140,14 @@ recipient := createTestRecipient()
 - Validation errors (400)
 - Invalid email format (400)
 - Invalid JSON (400)
+- Queue publish failure doesn't fail request
+
+### Edge Cases
+
+- Unicode and special characters (Chinese, emoji, umlauts)
+- Max length boundary validation (name: 255, subject: 500, message: 10000)
+- Empty honeypot is not detected as spam
+- Malicious content stored as-is (SQL injection, XSS, HTML, null bytes)
 
 ## API Characteristics
 
@@ -98,6 +156,7 @@ Messaging-api handles contact form and recipient operations:
 - **Contact Form**: Public endpoint for submissions with honeypot spam detection
 - **Messages**: Admin-only read access (GetAll, GetByID)
 - **Recipients**: Admin-only CRUD for email notification recipients
+- **Queue Integration**: Messages published to RabbitMQ for async processing
 
 ## Spam Protection
 
@@ -105,6 +164,7 @@ The honeypot test (`TestCreateContactMessage_SpamDetected`) verifies:
 
 - Messages with honeypot field filled return 201 (to not alert bots)
 - But the message is NOT saved to the database
+- And the message is NOT published to the queue
 
 ```go
 // Honeypot field (website) is filled - indicates spam
@@ -112,10 +172,19 @@ body := `{"name":"Bot","email":"bot@spam.com","subject":"Spam","message":"Buy no
 // Should return success but NOT save
 ```
 
+## Security Testing
+
+The `TestCreateContactMessage_MaliciousContent` test verifies:
+
+- SQL injection attempts are stored safely (GORM parameterizes queries)
+- XSS payloads are stored as-is (escaping happens at render time)
+- HTML content is preserved for legitimate use cases
+- Control characters and null bytes are handled
+
 ## Contributing Tests
 
-1. Follow naming: `Test<HandlerName>_<Scenario>`
-2. Organize by entity with section markers
+1. Add tests to the appropriate file matching the source file being tested
+2. Follow naming: `Test<HandlerName>_<Scenario>`
 3. Use table-driven tests for multiple scenarios
-4. Mock only the repository methods needed
+4. Mock only the repository/publisher methods needed
 5. Verify: `go test -cover ./internal/handlers/`
