+++ go.mod
@@ -3,7 +3,7 @@ module github.com/GunarsK-portfolio/messaging-api
 go 1.25.3
 
 require (
-	github.com/GunarsK-portfolio/portfolio-common v0.31.0
+	github.com/GunarsK-portfolio/portfolio-common v0.32.0
 	github.com/gin-gonic/gin v1.11.0
 	github.com/go-playground/validator/v10 v10.28.0
 	github.com/prometheus/client_golang v1.23.2
